Deep copy section children during deep copy of its root
Summary:
We currently don't do a proper deep copy of a `Section` during `ChangeSet` computation, and we think there are unexplained bugs and crashes because of it.

This diff will apply a deep copy of the entire tree. The change will be gated and not opened to all so that we can measure its effects.

Reviewed By: mihaelao

Differential Revision: D15489538

fbshipit-source-id: 60d8b8a0696e5573dc94e1293b966a2e5ebf6d00