Release component tree holders in RecyclerBinder when they're removed or detached.
Summary:
Release the ComponentTree when the items are removed in RecyclerBinder.

- For removeItemAt() / removeRangeAt() / replaceAll()
When items are removed, the corresponding views might want to disappear with animations, therefore posting a runnable to release the ComponentTrees later.

- For detach()
RecyclerBinder#detach() is triggered from RecyclerCollectionComponentSpec OnDetached method, it might be called on main or background thread. Because ComponentTree#release() can only be called on main thread, the trees is released immediately if we're on main thread, or post a runnable on main thread.

The change is protected by flag `ComponentsConfiguration.isReleaseComponentTreeInRecyclerBinder` to setup an experiment later.

Reviewed By: muraziz

Differential Revision: D15562548

fbshipit-source-id: d0aaee8af8d9f205cc69a223ed9b9a7b74037840