Add window-specific width in DoubleMeasureFixUtil class
Summary:
Replacing the width in DoubleMeasureFixUtil to be based on default display from WindowManager instead of being based on the width from the DisplayMetrics. The difference is that the dimensions from DisplayMetrics are specific to the entire screen size instead of just the window an app is running in. This is especially a problem for the layout of Messenger Kids in Chromebooks with content overflowing to the right of the window (since window width is smaller than screen width). See this Litho post (https://fb.workplace.com/groups/1430200360634661/permalink/2344121692575852) for more context.

I'm also updating all the callsites for correctWidthSpecForAndroidDoubleMeasureBug to include the WindowManager param. If the apps using those callsites are not being released on Chromebook then these changes ***__should__*** not affect those apps. However, I'm including the original ICs for those callsites as reviewers to ensure that this is actually the case.

Reviewed By: muraziz

Differential Revision: D15633534

fbshipit-source-id: 75d64a8a2ce1d1260d4b778c2ea3c8292f2f3dbc