Fix CachedValue bug when OnCalculateCachedValue has no params
Summary:
When no param is passed to method annotated with `OnCalculateCachedValue` we were not generating `hashCode` for the equivalent input class.
Some context about how `CachedValue` works: We generate a plain data class for each method annotated with `OnCalculateCachedValue` that encapsulates params and use that as key for cachedValues map in `ComponentTree`. If there was no param, we were skipping generating `hashCode`, therefore we couldn't retrieve cached value in other lifecycle methods, therefore always recreating new one.

Fix is to generate hashcode with class type when method has no param.

Reviewed By: topwu

Differential Revision: D15781618

fbshipit-source-id: 656bd7d81f006aa80f98e2d4a07a19babfdebce1