Do not assume that ComponentTree is available for every layout calculation
Summary:
When doing layout calculation we shouldn't assume that `ComponentTree` is always available and layout calculation should be independent from other pieces like `ComponentTree` or `LithoView`. Unfortunately `CachedValue` was breaking that assumption and creating hard dependency for `LayoutState` on `ComponentTree`. Null check here is not just duct taping, but it delivers correct behavior. That is, when `ComponentTree` is `null` then `OnCalculateCachedValue` is always called and returned (as opposed to returning existing value which doesn't exist when no ComponentTree is present).

One example where this was breaking is when you call `Component.willRender(c, component)` which shouldn't depend on `ComponentTree` existence, but breaks when component has `CachedValue`. With this diff, that no longer breaks.

Reviewed By: marco-cova

Differential Revision: D15803997

fbshipit-source-id: fe10864857e9a102263acd0e39171f57249a4ebc