Fix incorrect padding when customEllipsisText is enabled in TextSpec
Summary:
This's an issue reported from https://fb.workplace.com/groups/1430200360634661/permalink/2355095041478517/

When `customEllipsisText` is enabled, there's an extra padding at the end of the text. The root cause is that we use `Layout#getLineMax()` to calculate text width, but apparently the value is smaller than the actual width.

This diff just simply replace `Layout#getLineMax()` with existing `layoutWidth` which has been calculated in caller method.

Reviewed By: adityasharat

Differential Revision: D15874912

fbshipit-source-id: 3c1b6bd69e0e31af84c3093292d068b80541021a