Resolve RV padding ahead of time on main thread in RecyclerBinder
Summary: In the attached task, we crash because we try to resolve the padding for the first time from a BG thread. This solution is a hack, but much quicker to do than the alternative, which is to somehow pass padding into the RB instead of taking it off the View. Note, we'll probably need to do this in the long term for async changesets since we may kick off layouts before the RV is mounted and we can get padding from it.

Reviewed By: muraziz

Differential Revision: D15940228

fbshipit-source-id: 0233df9f39f93598a1e9dea280537640cc121aef