Allow RecyclerConfiguration to be copied
Summary:
We hope to clone a given `RecyclerConfiguration` and then overwrite some of its other values such as a different `RecyclerBinderConfiguration`.

Today `RecyclerConfiguration` is a read-only file, so I propose returning a `Builder` with all the default values copied from the original configuration. I will then be able to overwrite the original `RecyclerBinderConfiguration` before building a new config.

Reviewed By: mihaelao

Differential Revision: D15935044

fbshipit-source-id: 85614bc5cd34c892c992a2b9d95ee710c07dced2