Interface for listening to changeset events
Summary:
This adds a listener interface with a method that is invoked during onDataBound, after a changeset is applied.
The listener will be implemented by the plugin, which handles all the logic of parsing the data received and sending it to the desktop client.
If the listener is not configured, there's no behaviour change.
There are some helper methods which appear unused in ChangesetDebugConfiguration, but I want to land everything in one go so I can do a Litho release, so that the interface is accessible in the gradle plugin project.

Reviewed By: adityasharat

Differential Revision: D16120818

fbshipit-source-id: eec066e97029ef5d4bfc0ac2fd7bc4d8f336da0f