Reset child counter for Section copy
Summary: When creating a shallow copy for a component to use in a new changeset calculation, as a result of a state update, we need to clear the child counters of the cloned copy. Otherwise, when assigning unique global key to sibling Sections of the same type they will end up with different keys compared to the previous SectionTree, which makes diffing very inefficient.

Reviewed By: astreet

Differential Revision: D16204009

fbshipit-source-id: 8aa566700ae559d15ec0ef0a6c6d5da803332830