Set padding on LayoutManager to prevent race condition crash
Summary:
Fix for the attached crash. The main goal is to not access RV padding off the main thread as it will spuriously crash due to race conditions (as is probably expected).

The hacky solution here is to just cache the padding values on the LayoutManager once we have them. This essentially assumes they never change but for the most part this should be fine. (knock on wood)

Reviewed By: muraziz

Differential Revision: D16260309

fbshipit-source-id: 6f918ac608b2cfbf1b65d358d3eb18fa8ac8d685