Append the root component of LayoutSpecs to the InternalNode's components list.
Summary:
## ASpec.java
```
OnCreateLayout
static Component onCreateLayout(ComponentContext c) {
  return Row.create(c)
    .child(
      B.create(c)
    )
    .build();
}
```

## BSpec.java
```
OnCreateLayout
static Component onCreateLayout(ComponentContext c) {
  return Wrapper.create(c)
    .delegate(
      Text.create(c).text("text").build()
    )
    .build();
}
```

## Component hierarchy with their global keys

```

        A       {A}
        |
       Row      {A,Row}
        |
        B       {A,Row,B}
        |
     Wrapper    {A,Row,B,Wrapper}
        |
       Text     {A,Row,B,Wrapper,Text}

```

## What we see in Flipper

{F160975680}

**Where did Row and Wrapper go?**

Let's look at the `InternalNode` tree.

## InternalNode tree

```

        N1 [A]
         \
          N2 [Text, B]

```

Row and Wrapper were not added to the `InternalNode`'s components list because they are root components of the spec. Also notice that the global keys are still generated with the missing components.

## The problem

The root components are never added to the `InternalNode` tree; therefore we can never regenerate the same component hierarchy again (read *"never regenerate the same global keys again"*). Reconciliation works with the current implementation because coincidently we do not override global keys. For somewhat different reasons nested tree resolution also just works.

**We should fix this inconsistency because this is a risk for reconciliation, especially for prop update reconciliation and reusing cached measurements.**

## The obvious fix

Add the missing components to the `InternalNode` tree.

## The problem with the obvious fix

This breaks the current behavior of our internal API `InternalNode#getTailComponent()`. This used to be called `getRootComponent()` so let's use this moving forward.

If we add the root component of a `LayoutSpec` to the `InternalNode` it will change the return value of `InternalNode#getRootComponent` throughout Litho.

Let's look a `ASpec` again:

### ASpec.java

```
OnCreateLayout
static Component onCreateLayout(ComponentContext c) {
  return Row.create(c)
    .child(
      B.create(c)
    )
    .build();
}
```

This `LayoutSpec` will create an `InternalNode`; with our current implementation when we call `InternalNode#getRootComponent()` on its `InternalNode` we would get *(as expected)* component `A`. For `BSpec` it is broken, because its `InternalNode`'s root component will be `Text` instead of *(the expected)* component `B` because of completely different issue. *(which we shouldn't discuss right now)*

Now if we add the root component of `A` to the `InternalNode` then `InternalNode#getRootComponent()` would return `Row` instead. This would break some assertions throughout the framework. *(i.e. where ever `getRootComponent` is used instead of getting the component we would get the root component)*

This needs to be test, hence this behind an experiment config.

Reviewed By: astreet

Differential Revision: D16337152

fbshipit-source-id: 86fc9d6bd4469b8eac5e529c1f7d98d395ead38f