Add sub-adapter mode for RecyclerBinder
Summary: This mode can be used in cases where the RecyclerBinder is only controlling part of a list. In this case, mount/unmount should not be called (since the RecyclerBinder isn't in control of a RecyclerView), and instead the RecyclerView.Adapter should be retrieved with `getInternalAdapter` and placed within a multiplexing adapter (e.g. one that can present a consistent interface to the RecyclerView, but is backed by a list of Adapters).

Reviewed By: muraziz

Differential Revision: D16541401

fbshipit-source-id: 6d2f1411885b110e7dfb6cb613813f4a71cfd7d8