Back out "[Litho] Spannable leak fix"
Summary:
Original commit changeset: 51af4298e8e4

On devices <= 6.x, T48144659 a crash appears when typing in the search bar in MK sticker plugin.

The diff D16422508 attempted to address a related issue, but is causing problems on these devices, so backing this out for now.

This is a launch-blocking crash on Messenger Kids v85.

Reviewed By: Katalune

Differential Revision: D16611049

fbshipit-source-id: 3dd533b731b802a7d4b6fe051a9b438644a19c69