Allow manual viewport estimation to avoid sync initRange calls
Summary:
The goal here is to have a manual bypass around sync initRange calls on the main thread: we do this by specifying a manual viewport range so that a range doesn't have to be determined before kicking off async layouts.

Will follow up with Mihaela on longer term mitigation here (there are already params in flight).

Reviewed By: muraziz, paveldudka

Differential Revision: D16599558

fbshipit-source-id: 018a15a2aa986f5aef4235529c2a76990ab37c13