Added logging debounce.
Summary:
Added timeout to the logging:

file saving event - this event is happening too often, we reduce logging to once/hour for indication if saving is still triggered for the plugin,

navigation event - it is triggered when user is hovering over the symbol, and we don't want to send more than 1 navigation event for 1 click event.

error annotation event - it could be triggered few times per one error, and we want more hight-level tracking.

Reviewed By: muraziz

Differential Revision: D16656021

fbshipit-source-id: 6163cd50cf1440e4e77209e998e947858170bb46