Avoid remeasuring RecyclerSpec with wrapContent if bounds did not change after data change
Summary:
`wrapContent` feature of `RecyclerSpec` works by remeasuring the content (`RecyclerView`) whenever data has changed. This is to make sure to expand/collapse the orientation size as needed. Remeasuring is performed by triggering `RemeasureEvent` which will trigger state update which forces to remeasure the content.

There are cases where some state update triggers loop of remeasurement state updates when, for example, data model changes during onCreateLayout and updated model is passed to `DataDiffSection` which triggers remeasurement which calls onCreateLayout and it loops infinitely.

The solution is to make sure to avoid remeasurement when it's not needed. Whenever there is data change (`notifyDataSetChanged` is called) we fill viewport with new data and see whether the size of content changes and only when it changes we trigger `RemeasureEvent`.

Reviewed By: topwu

Differential Revision: D16715302

fbshipit-source-id: 1c118a4b9a8eaf911c621a38e752dd232f55cf41