Simplify handling of list of transitions from OnUpdateStateWithTransition
Summary:
Before we were collecting list of transitions in `StateContainer` and then consuming the list from it and adding to pending list of transitions. With this diff we collect transition right after calling `applyStateUpdate`, therefore the lifetime of transition in `StateContainer` is drastically reduced and instead the list is handled in `StateHandler`.

Even better approach would be to return `Transition` from `applyStateUpdate`, however, `StateHandler`/`StateContainer` is used not only for components but also for sections where there is no notion of `Transition`s. Therefore, transition handling in `StateContainer` is left as separate thing.

Note that with D16649349 ( 46314c537a67c3b0cc1ee7d8ffeb2c7cd1efdfae ) we make sure to use new instance of `StateContainer` when doing shallow copy, therefore, there is no longer need for synchronization within container.

Reviewed By: astreet

Differential Revision: D16731355

fbshipit-source-id: a112ec15f4f24be88a2b71c928e571f43ba4ca1f