Expose RecyclerBinder's Commit Policy through DynamicConfig
Summary:
# Problem

Today we don't have the ability to modify any configurations we set to the `RecyclerBinder` once they have been initialized. We don't want to allow all the configurations to be mutable, but `rangeRatio` and `commitPolicy` should be allowed to change throughout a production's interaction.

# Solution
Expose the configurations we allow to change through `DynamicConfig`. Today I'm only going to allow `commitPolicy` to be modified.

I recommend that the configuration changed to be exposed through `RecyclerCollectionComponent` through `Trigger`.

Reviewed By: astreet

Differential Revision: D16674728

fbshipit-source-id: 89277591adddcaae9dc23fcbf905c3e069410fd9