Invalidate measurement in RecyclerBinder when initial measurement was performed with empty data
Summary:
In D16715302 (1c118a4b9a8eaf911c621a38e752dd232f55cf41) I changed RecyclerBinder's remeasurement logic so that it avoids remeasuring if the new dataset doesn't change the RecyclerView's bounds.

However, there are cases where we should not avoiding remeasurement and not rely on previous measurement. Imagine a scenario where the RecyclerView was measured before any dataset was bound to it which results in one of the dimensions being 0. When new dataset is bound fillviewport will try to fill the main orientation direction and may result in the same sizes as previous measurement (one side being 0). We shouldn't avoid remeasurement in those cases.

Reviewed By: topwu

Differential Revision: D16805306

fbshipit-source-id: 6204f34a00c0af01e66cd3e2f8c5b073365a7c6e