Fixed inconsistency on Component file update
Summary:
When we run build compile and after that updated Component with the plugin, file changes on the disc were inconsistent with the Document updated by the plugin. Now we update file instead of a Document to fix this.

Document represents the contents of a text file loaded into memory, and possibly opened in an IDEA text editor.
A PSI (Program Structure Interface) file represents the contents of a file as a hierarchy of elements, and can be saved to disk.

To make Document consistent with the file system I called FileDocumentManager.getInstance().saveDocument(componentDocument); but that didn't help.

Reviewed By: muraziz

Differential Revision: D16808754

fbshipit-source-id: a7dc6ef9c4f8571d6c58526a7d7489cb818dc395