Added wildcard support for guessing class name
Summary: In IntelliJ environment we are creating a Class Name object. We use ClassName.bestGuess(typeName) method, but it handles only the valid qualified Java names package.Name and throws an error for the generics or wildcards. That's why we are working around this by excluding such elements from the method name.

Reviewed By: astreet

Differential Revision: D16858982

fbshipit-source-id: 901f3dcd85365c8f79f550129b8516563fa127ad