Add config to applyReadyBatches before mounting to RecyclerView
Summary: In cases where we are mounting H-Scrolls, it's possible that the parent RecyclerView blocked on their layout/section computation via a LayoutState future, but the runnable to update the adapter on the main thread hasn't run -- this gives a chance to update the adapter before we attach to this RecyclerView. Testing this in a few places.

Reviewed By: topwu

Differential Revision: D16970934

fbshipit-source-id: 3f77abba6944a437fbb2a267eedf2914cfc43531