BREAKING: Always measure component inside a LayoutState calculation
Summary:
went through calls of Component.measure that were made from outside a Spec class and replaced them with createComponentTreeAndMeasure, which creates a ComponentTree to measure the component inside a LayoutState calculation.
The util methods that are calling measure could still be called from a Spec during a layout so we delegate to measure() if that's the case.

Reviewed By: pasqualeanatriello

Differential Revision: D14458805

fbshipit-source-id: b5dc8f3b61966a3aa7da1abdac9f166e534d363d