Keep cached layout on LayoutState
Summary:
We want to keep the cached layouts in the LayoutState instead keeping them on the Component instance, because this guarantees that a cached InternalNode would only be accessed on a single thread (since LayoutStates are bound to a single thread).
This diff moves the caching to the LayoutState. The caching tests that used a measured component to verify that layout caching works had to be changed to reuse the same LayoutState instance to verify caching. That's why some tests had to be moved to the powermock package: we needed to static mock ComponentsPools so that we can force the same layoutState instance to be returned when acquiring and be able to verify caching.

Reviewed By: astreet

Differential Revision: D13528364

fbshipit-source-id: 2770b7799d090419e7b33001676586a2ae4f6b3d