Immediately remove MountItem mapping in unmountItem to protect against re-entrancy
Summary: This is likely the cause of the attached crash: I'm not sure exactly where the re-entrancy is coming from and I'd have to add more logging to figure it out. However this change just seems more correct in general since we will now never be able to unmount an item while it is in the middle of being unmounted.

Reviewed By: muraziz

Differential Revision: D17158727

fbshipit-source-id: df93ed94c71baf0123b66ae2055ebb49af517b54