Don't clear LayoutState from ComponentContext if layout is interrupted
Summary:
When a layout calculation is interrupted, the partial InternalNode still holds on to the ComponentContext which has the current LayoutState reference and which will be used to measure the component.
We should only clear the LayoutState ref from the ComponentContext when the layout calculation is done and not if there was an interruption.

Reviewed By: marco-cova

Differential Revision: D17206249

fbshipit-source-id: aff5aa77e735fa10298e051fca6e246ae03bc559