Replace internal uses of ComponentsLogger#emitMessage to ComponentsReporter#emitMessage
Summary:
`ComponentsLogger` should be doing only perf logging and error reporting should be handled by `ComponentReporter`. Currently it is a mix of both and this diff replaces usages of `ComponentsLogger`'s  error reporting APIs with corresponding APIs in `ComponentsReporter`.

The implications of this change is that soft errors (LogLevel.ERROR) will be reported more often (https://fburl.com/diffusion/i2cc5gtp) as opposed to reporting error only when ComponentLogger instance is provided via `ComponentContext`. I also checked that we don't have LogLevel.FATAL cases which would crash the app, therefore, the effect of this change would be minimal.

Reviewed By: topwu

Differential Revision: D17158825

fbshipit-source-id: 1f1597ff4e8c827016e316aa86a47464ce62d139