Move setting ItemAnimator to mount/unmount
Summary:
Resetting the ItemAnimator in onBind onUnbind can cause animations to be interrupted when the Recycler component receives an update even if this update would not cause an unmount/mount cycle to happen.
This diff moves the ItemAnimator to mount/unmount instead. Since comparing ItemAnimator instances in shouldUpdate is not practical and would fail for the system provided animators, we compare just the type assuming that animators of the same type would animate items in the RecyclerView in the same way.

Reviewed By: adityasharat

Differential Revision: D17163032

fbshipit-source-id: f8a83185b905d18974f651688c949defeba68a6d