Stop and clean running transitions not existing nor declared in the new layout
Summary:
When we have a running transition and we compute the new Layout, the transition items are locked and won't be unmounted until the transition finishes even if there's no disappearing transition declared.

With this diff, we want to collect all transition which disappeared without a disappearing transition defined. Then we check if they are still running and stop and releasing them to be unmounted.

Reviewed By: astreet

Differential Revision: D17261342

fbshipit-source-id: 77cb1648c0bd1b00ee3ec8bc3d16ab3dfe782bca