Delay applyReadyBatches if RecyclerView is computing layout
Summary:
We can't apply async batches on the adapter while RecyclerView is scrolling. There's currently a check to delay applying operations until RecyclerView is done computing layout, but in subadapter mode that gets ignored because the RecyclerBinder doesn't have access to the RecyclerView.
This provides a hook into the RecyclerView which is used in subadapter mode and does the same check as we do for a mounted RecyclerView.

Reviewed By: astreet

Differential Revision: D17366662

fbshipit-source-id: eec8d6208492323394cedfd479005c0117c42cf3