Experimental feature in Sections to enable posting the first ChangeSets to the front of the UI queue
Summary:
# Problem
When the first layout is done on a background thread, the changeset is posted to the back of UI thread queue for `RecyclerBinder` to be handled.

This means that the first UI runloop is going to mount the `Recycler` without any data. This results in a momentarily blank UI before the loaded state is shown.

# Solution
This diff enables us to experiment with posting the changeset to the front of the queue when it's the first changeset computed on the background thread.

We won't have correctness issue here because if `useBackgroundChangeSet` is enabled, we will never attempt to post the changeset to the UI thread.

Reviewed By: SimoneCasagranda, mihaelao

Differential Revision: D17349183

fbshipit-source-id: af26298088c92da68217cee1f09742a51d54ee2e