Don't sync isMeasured()
Summary:
Removing `synchronized` access from isMeasured(). `mIsMeasured` and `mRequiredRemeasure` are only written to in the same synchronized block during measure() - we can remove this without any implications.
Needed to start experiment with non-blocking init range - with the methos synchronized, the first async layout still waits for measure to finish before moving on to compute the next async layout, because it calls into isMeasured() and waits for the RecyclerBinder lock to be freed.

Reviewed By: pasqualeanatriello

Differential Revision: D17451104

fbshipit-source-id: c1359139600dc1130930735e7d7582ed91649abf