Background async layouts don't need to wait on LayoutStateFutures
Summary:
LayoutStateFutures wait on other compatible LayoutStateFutures to get a LayoutState result. A LayoutStateFuture created on a background thread could also get blocked waiting on another LayoutStateFuture - in the case where the layout it's computing was an async operation and it doesn't need to immediately return a size this is unnecessary.
pasqualeanatriello astreet can you think of any cases when we do want the bg thread to wait for the result from another LayoutStateFuture?

Reviewed By: astreet

Differential Revision: D17346759

fbshipit-source-id: 05e649fd99312d7c5a5554e34b2155899056e9d0