Correctly remove PostDispatchDrawListener when SectionAdapter is detached from RecyclerView
Summary:
This is the follow-up of D17318082, which causes NullPointerException crash as described in T54393283.

In the past, when RecyclerBinder call `setPostDispatchDrawListener(null)` to remove the listener, BetterRecyclerView didn't remove it but set the field to null, which causes the crash next time `dispatchDraw()` is called.

The diff separate `setPostDispatchDrawListener` to `registerPostDispatchDrawListener` and `unregisterPostDispatchDrawListener`, and correctly remove the lister when it's no needed.

Reviewed By: mihaelao

Differential Revision: D17526055

fbshipit-source-id: 0ea199c764aadde703dfa74f0652a27b5179e45a