Allow TreeProps to be used in OnCreateInitialState of Sections
Summary:
# Problem

`TreeProps` are special props that come from the parents of a section in its hierarchy, and it is expected to be accessible in the `OnCreateInitialState` function of the child.

However, that's not the case today, and it actually comes as a surprise because there should be no restriction around where should `TreeProps` be accessible in the child section.

# Solution

We can populate the `TreeProps` in the child section earlier in the `createNewTreeAndApplyStateUpdates` function of `SectionTree`. This allows us to use `TreeProps` in `OnCreateInitialState`.

Reviewed By: mihaelao

Differential Revision: D17562805

fbshipit-source-id: 90bd9be56f54ddb07427af0a510b393b277ecd4c