Create ComponentContext with StateHandler before measuring outside a LayoutState
Summary:
If a Component which has state on it is being measured outside a LayoutState it would cause a NPE when calculating state.
An example where this is happening now is described here: https://fb.workplace.com/groups/1430200360634661/permalink/2433330843654936/
A stateful component is measured in a Section with a SectionContext - which always has a null StateHandler. This will cause a crash if calling `measure`, but we can prevent this from `measureMightNotCacheInternalNode` for existing callsites.

Reviewed By: adityasharat

Differential Revision: D17571594

fbshipit-source-id: ef5ff7031eb9f4c302d46e76e22c527d400daf7b