Don't override LayoutState.isPartialLayoutState after it's set to true
Summary:
Potentially the cause of a thread safety issue as follows:
1. BG thread is running a LayoutStateFuture
2. UI thread interrupts the Bg thread LSF
3. When interrupted, BG thread marks LayoutState.mIsPartialLayoutState = true
4. UI and BG thread both return from waiting for result here: https://fburl.com/diffusion/1lgp8u0e
5. UI thread resumes and finishes the calculation and marks LayoutState.mIsPartialLayoutState = false
6. BG thread reads result.isPartialLayoutState() here: resolvePartialInternalNodeAndCalculateLayout as false, even though it was interrupted and should return null
We can end up using the same LayoutState instance on two different threads which should not happen.
Overriding the flag is unnecessary too, all threads that can read it after it's been set actually only need to know that the layoutState was interrupted at some point.

Reviewed By: marco-cova

Differential Revision: D17573176

fbshipit-source-id: 3ae52684c2d7642e081d4bd556cca1a64962d95c