Shallow copies the tail component of the nested tree before resolving it.
Summary:
Shallow copies the tail component of the nested tree before resolving it. In addition set the global key on the shallow copy. This will ensure that global keys for nested trees remain stable.

Consider the component hierarchy:
```

A
 \
  B(with-size-spec)
   \
    C

```

The expected global key for C is `A,B,C` but actual will be `A,B!1,C`. This will break application of state update during reconciliation.

In fact during the first layout pass the global key for will be `A,B,C`; but when is B (which is a nested tree) is resolved during nested tree resolution the global keys get regenerated to `A,B!1,C` because `A` has its child counter for `B` set at 1.

Differential Revision: D17630411

fbshipit-source-id: 2d770ebe1e8c3b8f7894e7a607834fc08b419903