Move LayoutStateFuture from ComponentContext to LayoutState
Summary:
Right now we pass in a LayoutStateFuture down the tree through the ComponentContext to check if a bg layout needs to be interrupted and resumed on the UI thread.
Each ComponentContext created as part of the layout state calculation holds a reference to this LayoutStateFuture without ever removing it. This leads to memory leaks as seen in the hprofs below.

This changes how the LayoutStateFuture is propagated down the tree - it's now saved in the LayoutStateReferenceWrapper which releases the reference to it at the end of the LayoutState calculation. This means that the LayoutStateFuture reference will only be around for the scope of the LayoutState calculation and released after.

In the next diff I'll rename LayoutStateReferenceWrapper to LayoutStateContext, as it's starting to take more shape towards that.

Reviewed By: pasqualeanatriello

Differential Revision: D17684813

fbshipit-source-id: 41cae240881f01fd17caeeec0ab57d92a8013eb3