ComponentWarmer: add entry for tag before layout is finished
Summary: we are already doing this for async, should be the same for sync. We want to add the entry before the layout is finished, that way the RecyclerBinder can add the correct ComponentTreeHolder to the adapter and be able to use LayoutStateFutures to wait for the ongoing layout instead of starting another from scratch.

Reviewed By: astreet

Differential Revision: D17715444

fbshipit-source-id: 547473c5834c241d4d11a453fb5cbeaa5eb86fe3