Postpone ComponentTree#mountComponent for reentrant mounts, then mount new layoutState afterward
Summary:
The diff is trying to propose a solution for T37535936(and probably T54518955 if it has same root cause). The root cause is the crash is because the component tries to update a new root synchronously in VisibleEvent handler, which causes MountState to mount a LayoutState when it's already mounting.

D17686710 tries to provide repro steps in our playground app. I'm working to provide a unit test, but it might take some time and it would help a lot if you can give me some feedback about the diff beforehand.

Reviewed By: astreet

Differential Revision: D17607247

fbshipit-source-id: 10bc040ecc043ae4082deed3515194987b3c1e13