Refactor the implementation of owner key for transitions
Summary:
This diff fixes transition animations for the use case defined in the test plan. The cause was the mismatch between the transition id generated for `InternalNode` (in `LayoutState#getTransitionIdForNode`) and the owner key set on the transition when it was created (in `TransitionUtils#setOwnerKey`).

This diff ensures that the transition id for the internal node is based on the component which creates the Transition for it.

The logic for `AUTOGENERATED` and `GLOBAL` transition key type remains unchanged. The following is the modified logic for generating the transition id for an InternalNode which has a local transition key set on it (i.e. `SCOPED`).

```
if InternalNode has transitions
     transition id key = the global key of the component which created the transition.
else
     transition id key = the global key of the owner of the tail component.
```

**owner key:** The owner key of an instance of component A is the global key of the component which created it. (i.e. called `Component.Builder#create(ComponentContext)` for it)

Reviewed By: astreet

Differential Revision: D17367458

fbshipit-source-id: fbb9ddd751618637530f9871734e005f33950ca2