refactor layout id calculation, fix type id size error
Summary:
Refactored the way the max shifts are defined. Included explicit bit allocations and erased hard-coded values.

Fixed bug where for a component of type `border`, the type id is `4=0b100`. However the layout ID only allocated 2 bits to the type field. Therefore the type attribute overflowed into the level attribute causing border elements to be rerendered when was unnecessary.

Updated the tests to match new format and added a border type test.

Reviewed By: astreet

Differential Revision: D17673003

fbshipit-source-id: 1b0522e24fe6826737720b17cc569ab4b73c4b90