Fix race condition when binding events and triggers in ComponentTree
Summary:
# Problem

With layout state future, both `ComponentTree.measure()` and `ComponentTree.calculateLayout()` will have access to the same instance of `LayoutState`. Both functions also do `ComponentTree.bindEventAndTriggerHandlers()` based on the availability of the list of `Components` in `LayoutState`.

This led to a race condition where if one of the functions processed it first and clear all the `Components` on `Thread A`, while the other on `Thread B` will assume that no `Components` are available and will unbind all registered handlers.

# Solution
The list of available `Components` is used to determine if `EventHandler` and `TriggerHandler` should be registered or cleared. Hence, we should only bind them once.

The proposal here is to do the binding and unbinding process only once, and if the other recognizes that the process has been completed, it will skip it entirely.

The binding process is skipped when the list is null, __not__ when it's empty.

Reviewed By: mihaelao

Differential Revision: D17806117

fbshipit-source-id: aee72d5b50262272705e078b8d20dd8595567682