Fix an NPE trying to start a delayed animation which was canceled
Summary:
It appears a delay transition tries to start a transitionId which doesn't exists anymore in the TransitionManager mAnimationStates. This could be due to Transition which got cancelled before even starting. However, for some reason the delay transition still calls shouldStart even if not active anymore.

This doesn't fix the root problem since I couldn't identify it. However, should solve the NPE.

Differential Revision: D17814603

fbshipit-source-id: 48e8e24576d137841622beba103b78dfa242746f