IOOB in ComponentWarmer: Remove ComponentTreeHolder from ComponentWarmer as soon as it's used once
Summary:
Here's what I *think* could be causing the IndexOutOfBounds exception in t55405758. astreet pasqualeanatriello if this makes sense I can try to write up a test, but wanted to check if it's reasonable to suspect this.

When we create a new AsyncInsertOperation for a ComponentTreeHolder on the async changeset thread, we call `setInserted(false)` on it.
Later when we apply an async insert operation for a ComponentTreeHolder on the UI thread, we call `setInserted` to true on it.

Let's say we have prepared a ComponentTreeHolder for position 0 and cached in the ComponentWarmer, and we have two changesets with the following operations:
Changeset 1: AsyncInsert 0
Changeset 2: AsyncRemove 0, AsyncInsert 0, AsyncInsert 1.

It looks like it's possible for Changeset1 to applyReadyBatches and change ComponentTreeHolder#inserted to true on the UI thread right after we created the AsyncInsertOperation for Changeset 2 and set ComponentTreeHolder#inserted to false. That means that when Changeset2 wants to applyReadyBatches, we could skip inserting the item at position 0 because this check fails: https://fburl.com/diffusion/zrhkwxj2. Since we've also just removed item at position 0, when inserting at position 1 this could cause that crash.
We could prevent this by making sure we only use a cached ComponentTreeHolder once

Reviewed By: astreet

Differential Revision: D17951320

fbshipit-source-id: 0c22a4354c2e83c2d7b55c7dd964a31080ef9629