Cache device type in DoubleMeasureFixUtil
Summary:
The hasSystemFeature call in the package manager (used here to determine if this device is a Chromebook) is measurably expensive, and since we can reasonably assume that the device type will not change once the application is installed, it makes sense to cache the result of the first lookup to improve runtime performance.

Preliminary results show a large improvement over the course of several thousand calls.

Reviewed By: stashuk

Differential Revision: D17802064

fbshipit-source-id: 59a8ece6bca065d778bd02b83de8d11c57e504b6