Add Litho ComponentsConfiguration and gating for measuring efficacy of device cache change in DoubleMeasureFixUtil.
Summary:
The hasSystemFeature call in the package manager (used here to determine if this device is a Chromebook) is measurably expensive, and since we can reasonably assume that the device type will not change once the application is installed, it makes sense to cache the result of the first lookup to improve runtime performance.

Preliminary results show a large improvement over the course of several thousand calls.

This change adds experimental gating to the changes included in "Cache device type in DoubleMeasureFixUtil" (D17802064).

Reviewed By: stashuk

Differential Revision: D17828272

fbshipit-source-id: 83b29ef669c58ccf09faa0d4aee4530dff6fbc03