Narrow type constructor parameter to int.
Summary:
Previously this constructor parameter took any `Object` but was used
only to accept `int`s.  This diff tweaks the constructors of
`ComponentLifecycle` such that the default (no-args) constructor
always associates the Type ID it creates with the instance's class,
and the constructor that accepts a type parameter always expects a
custom one (it no longer supports a sentinel value to indicate that it
should take the default value of `getClass()`), which has to be a
primitive `int`.

Reviewed By: marco-cova

Differential Revision: D17857294

fbshipit-source-id: bae4e3d0c6467e6ee37a19884e8018c54d422b5b