Disable createInitialState once when layout cancellation is enabled
Summary:
In the warmed layouts cancellation experiment, we see some cases when onCreateInitialState is not applied.
When we ran the cancellation experiment previously we tested it along another feature which was enforcing onCreateInitialState to be called once and only once even if there were multiple threads computing the first layout simultaneously.
This was silently enabled for the nf+sections cancellation experiment but it's buggy and the problem this feature was trying to fix should never occur in this case.
I'll hardcode disabling this feature in this diff so we can pick it and follow up with proper cleanup.

Here's what I think is happening:
- we cancel a layout but still have a child resolving, which is trying to call onCreateInitialState
- when we triggered the cancellation we've released the ComponentTree here: https://fburl.com/diffusion/vilx6gyq, which also nulls the ComponentTree's StateHandler instance
- when we go and try to commit the initial state value to the ComponentTree in the new feature codepath, the state handler is null so the initial state value is not commited: https://fburl.com/diffusion/7trzaqkg
- we then go on and createLayout for the component, which will crash when reading the initial state value

Differential Revision: D18227109

fbshipit-source-id: 977994cd73eb108375f0397c49eface175075ef1